fix cr issue 18 31
