+++ AElf.All.sln
@@ -383,6 +383,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AElf.Kernel.SmartContract.O
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AElf.Silo.Launcher", "src\AElf.Silo.Launcher\AElf.Silo.Launcher.csproj", "{81AFEFAB-F4C5-4F1D-85F5-FADBB3EB1ACF}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AElf.Kernel.SmartContract.Grain", "src\AElf.Kernel.SmartContract.Grain\AElf.Kernel.SmartContract.Grain.csproj", "{45C36EAC-D5F2-4AB0-9526-F6DC4CECBD2C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -1101,6 +1103,10 @@ Global
 		{81AFEFAB-F4C5-4F1D-85F5-FADBB3EB1ACF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{81AFEFAB-F4C5-4F1D-85F5-FADBB3EB1ACF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{81AFEFAB-F4C5-4F1D-85F5-FADBB3EB1ACF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{45C36EAC-D5F2-4AB0-9526-F6DC4CECBD2C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{45C36EAC-D5F2-4AB0-9526-F6DC4CECBD2C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{45C36EAC-D5F2-4AB0-9526-F6DC4CECBD2C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{45C36EAC-D5F2-4AB0-9526-F6DC4CECBD2C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1294,5 +1300,6 @@ Global
 		{6981684D-450C-412B-9082-0B0A67A679B9} = {D3950CC9-808F-4ED8-946A-79A992F3F8EF}
 		{0DBE8AFD-19AA-469A-A1B5-8939CF8A0606} = {90B310B4-C2DB-419E-B5EE-97FA096B62CC}
 		{81AFEFAB-F4C5-4F1D-85F5-FADBB3EB1ACF} = {798BAA50-825B-4DD9-83E2-77A9EE6EBE09}
+		{45C36EAC-D5F2-4AB0-9526-F6DC4CECBD2C} = {90B310B4-C2DB-419E-B5EE-97FA096B62CC}
 	EndGlobalSection
 EndGlobal
