+++ src/AElf.Kernel.SmartContract.Grain/AElf.Kernel.SmartContract.Grain.csproj
@@ -1,0 +1,30 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <RootNamespace>AElf.Kernel.SmartContract.Orleans</RootNamespace>
+        <PackageId>AElf.Kernel.SmartContract.Grain</PackageId>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Microsoft.Orleans.CodeGenerator.MSBuild" Version="3.7.1">
+            <PrivateAssets>all</PrivateAssets>
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        </PackageReference>
+        <PackageReference Include="Microsoft.Orleans.Core.Abstractions" Version="3.7.1" />
+        <PackageReference Include="MongoDB.Bson" Version="2.17.1" />
+        <PackageReference Include="Orleans.Providers.MongoDB" Version="3.8.0" />
+        <PackageReference Include="Volo.Abp.Autofac" Version="5.2.2" />
+        <PackageReference Include="Volo.Abp.Caching" Version="5.2.2" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\AElf.Kernel.SmartContract.Orleans\AElf.Kernel.SmartContract.Orleans.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <Folder Include="State" />
+    </ItemGroup>
+
+</Project>
