+++ src/AElf.Launcher/appsettings.json
@@ -49,7 +49,6 @@
   },
   "SiloOrStandalone": "Silo",
   "SiloInstanceCount": 2,
-  "LaunchProgram": "BP",
   "Orleans": {
     "ClusterId": "AELFSiloServerSiloClusterBP1",
     "ServiceId": "AELFSiloServerOrleansBasicServiceBP1",
