+++ src/AElf.Silo.Launcher/Extensions/OrleansHostExtensions.cs
@@ -61,29 +61,28 @@ public static class OrleansHostExtensions
                 })
                 .Configure<ClusterMembershipOptions>(options =>
                 {
-                    options.DeathVoteExpirationTimeout = TimeSpan.FromSeconds(1);
-                    options.ProbeTimeout = TimeSpan.FromSeconds(1);
+                    options.DeathVoteExpirationTimeout = TimeSpan.FromSeconds(Constants.DeathVoteExpirationTimeout);
+                    options.ProbeTimeout = TimeSpan.FromSeconds(Constants.ProbeTimeout);
                 })
                 .Configure<GrainCollectionOptions>(options =>
                 {
-                    options.DeactivationTimeout = TimeSpan.FromMinutes(1);
-                    options.CollectionAge = TimeSpan.FromMinutes(2);
-                    options.CollectionQuantum = TimeSpan.FromMinutes(1);
+                    options.DeactivationTimeout = TimeSpan.FromMinutes(Constants.DeactivationTimeout);
+                    options.CollectionAge = TimeSpan.FromMinutes(Constants.CollectionAge);
+                    options.CollectionQuantum = TimeSpan.FromMinutes(Constants.CollectionQuantum);
                 })
                 .Configure<PerformanceTuningOptions>(opt =>
                 {
-                    opt.MinDotNetThreadPoolSize = 20480;
-                    opt.MinIOThreadPoolSize = 200;
-                    opt.DefaultConnectionLimit = 200;
+                    opt.MinDotNetThreadPoolSize = Constants.MinDotNetThreadPoolSize;
+                    opt.MinIOThreadPoolSize = Constants.MinIOThreadPoolSize;
+                    opt.DefaultConnectionLimit = Constants.DefaultConnectionLimit;
                 })
                 .Configure<SchedulingOptions>(opt =>
                 {
-                    opt.MaxActiveThreads = 200;
+                    opt.MaxActiveThreads = Constants.MaxActiveThreads;
                 }).Configure<SiloMessagingOptions>(opt =>
                 {
-                    opt.ResponseTimeout = TimeSpan.FromSeconds(10);
+                    opt.ResponseTimeout = TimeSpan.FromSeconds(Constants.ResponseTimeout);
                 })
-                // .AddMemoryGrainStorage("PubSubStore")
                 .ConfigureApplicationParts(parts => parts.AddFromApplicationBaseDirectory())
                 .UseDashboard(options =>
                 {
