+++ src/AElf.Silo.Launcher/appsettings.json
@@ -18,7 +18,6 @@
       "Default": "Trace"
     }
   },
-  "LaunchProgram":"Silo",
   "Orleans": {
     "ClusterId": "AELFSiloServerSiloClusterBP1",
     "ServiceId": "AELFSiloServerOrleansBasicServiceBP1",
